refactor(progress indicator): removes Circular Progress component
DSY-1896 - Update | Progress Indicator
