+++ src/components/ProgressIndicator/ProgressIndicator.device.tsx
@@ -2,5 +2,4 @@ import { storiesOf } from '@storybook/react-native';
 
 import { all } from './ProgressIndicator.stories';
 
-storiesOf('ProgressIndicator', module)
-  .add('all', all);
+storiesOf('ProgressIndicator', module).add('all', all);
