+++ src/components/ProgressIndicator/ProgressIndicator.tsx
@@ -1,8 +1,11 @@
-import React from 'react';
+import React, { useEffect, useRef } from 'react';
 import { withTheme } from 'styled-components/native';
 import { Theme } from '@naturacosmeticos/natds-themes/react-native';
+import { Animated, Easing, Platform } from 'react-native';
+import {
+  Container, Layer, Line, Loop,
+} from './ProgressIndicator.styles';
 import { getSize } from '../../common/themeSelectors';
-import { CircularProgress } from './CircularProgress/CircularProgress';
 
 export type ProgressIndicatorSizes = 'standard' | 'semi' | 'medium' | 'large';
 
@@ -13,16 +16,76 @@ export interface ProgressIndicatorProps {
   size?: ProgressIndicatorSizes;
   /**
    * Default theme
-  */
+   */
   theme: Theme;
 }
 
 export const ProgressIndicatorComponent = ({
   size = 'medium',
   theme,
-}: ProgressIndicatorProps) => (
-  <CircularProgress size={getSize(theme, size)} />
-);
+}: ProgressIndicatorProps) => {
+  /**
+   * Duration specify how much the circle will take to make a 360deg loop around itself,
+   * decrease it will speed up the animation speed and increase will slow the animation speed
+   * The default speed is 1 second per loop
+   */
+  const duration = 1000;
+  /**
+   * This animation/Animated.timing, is responsible for looping the border around the view.
+   */
+  const timer = useRef(new Animated.Value(0)).current;
+  const rotation = Animated.timing(timer, {
+    duration,
+    easing: Easing.linear,
+    isInteraction: false,
+    toValue: 1,
+    useNativeDriver: Platform.OS !== 'web',
+  });
+  /**
+   * The rotate animation will take from 0deg to 360deg to make a full loop around itself
+   */
+  const fullCircularRange = 360;
+  const minCircularRange = '0deg';
+  const maxCircularRange = `${fullCircularRange}deg`;
+  const layerStyle = {
+    transform: [
+      {
+        rotate: timer.interpolate({
+          inputRange: [0, 1],
+          outputRange: [minCircularRange, maxCircularRange],
+        }),
+      },
+    ],
+  };
+
+  /**
+   * Loop rotation animation continuously,
+   * each time it reaches the end, it resets and begins again from the start.
+   */
+  function startRotation(): void {
+    timer.setValue(0);
+    Animated.loop(rotation).start();
+  }
+
+  /**
+   * Reset the timer and loop the animation again on each update
+   */
+  useEffect(() => {
+    startRotation();
+  }, []);
+
+  return (
+    <Layer as={Animated.View} size={getSize(theme, size)}>
+      <Loop as={Animated.View}>
+        <Layer as={Animated.View} size={getSize(theme, size)} style={layerStyle}>
+          <Container as={Animated.View} size={getSize(theme, size)}>
+            <Line as={Animated.View} size={getSize(theme, size)} />
+          </Container>
+        </Layer>
+      </Loop>
+    </Layer>
+  );
+};
 
 export const ProgressIndicator = React.memo(
   withTheme(ProgressIndicatorComponent),
